That the project aims to answer include:
What are the most common types of accidents that lead to insurance claims?
Which types of vehicles are most likely to be involved in accidents?
What is the average amount of the claim for different types of accidents?
Which locations have the highest number of insurance claims?
By completing this project, beginners will gain hands-on experience in using Tableau to analyze and visualize data related to vehicle insurance claims. This will enable them to develop their skills in data analysis and data visualization, which are critical in many industries today.